fix(manage): change loadsites cmd to update_site_default
